requirements.txt
# AI and API dependencies
anthropic>=0.25.0
openai>=1.0.0  # Fallback option
pillow>=10.0.0
requests>=2.31.0

# Home Assistant integration
homeassistant-api>=0.2.0
pyyaml>=6.0.1

# Logging and utilities
python-dotenv>=1.0.0
schedule>=1.2.0

# Development/testing
pytest>=7.4.0
python-dateutil>=2.8.2